pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dorsalhub"
version = "0.2.0"
description = "Dorsal is a local-first metadata generation and management toolkit."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Rio Achuzia", email = "rio@dorsalhub.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",    
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "blake3>=1.0.8",
    "langcodes[data]==3.5.0",
    "pydantic>=2.12.1",
    "python-dateutil>=2.9.0.post0",
    "python-magic>=0.4.27; sys_platform == 'linux' or sys_platform == 'darwin'",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "requests>=2.32.3,<3.0",
    "rich>=14.0.0",
    "textual>=6.2.1",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
    "jsonschema-rs>=0.38.1",
    "pypdfium2>=5.2.0",
    "pymediainfo>=6.1.0",
    "tomlkit>=0.12.0",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
tlsh = [
    "py-tlsh>=4.7.2"
]

pandas = [
    "pandas>=2.0.0"
]

all = [
    "dorsalhub[tlsh]",
    "dorsalhub[pandas]",
]

[project.urls]
"Homepage" = "https://dorsalhub.com"
"Repository" = "https://github.com/dorsalhub/dorsal"
"Documentation" = "https://docs.dorsalhub.com"

[project.scripts]
dorsal = "dorsal.cli:cli_app"

[tool.hatch.build.targets.wheel]
packages = ["src/dorsal"]
include = ["LICENSE", "NOTICE", "/src/dorsal/schemas/open/*.json"]

[tool.hatch.build.targets.sdist]
exclude = [
  "/reference",
  "/notebooks",
  "/scripts",
  "uv.lock",
  ".github",
]

[tool.pytest.ini_options]
log_cli = true
log_level = "DEBUG"
log_cli_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]

[tool.ruff]
extend-exclude = ["*.ipynb"]
line-length = 120

[tool.ruff.lint]
select = [ "E4", "E7", "E9", "F", "B" ]
ignore = ["F401"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F811"]

[tool.mypy]
plugins = "pydantic.mypy"
exclude = ["dist/", "build/", "venv/", ".venv/"]
mypy_path = "src"
warn_unreachable = true

[dependency-groups]
dev = [
    "black>=24.10.0",
    "mypy>=1.15.0",
    "pandas>=2.3.1",
    "pyfakefs>=5.9.2",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "requests-mock>=1.12.1",
    "ruff>=0.8.4",
    "types-python-dateutil>=2.9.0.20250809",
    "types-tqdm>=4.67.0.20250809",
]
